Console tuned

+++ w2/admin.php
@@ -171,7 +171,7 @@ class Admin
                 $result &= (int)unlink($fn);
         }
         global $sky;
-        $s = substr($sky->s_statp, 0, -1) + 1;
+        $s = (int)substr($sky->s_statp, 0, -1) + 1;
         $sky->s_statp = $s > 9999 ? '1000p' : $s . 'p';
         return $result;
     }
